+++ next.config.ts
@@ -3,9 +3,7 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   reactCompiler: true,
   output: "standalone",
-  module.exports = {
   allowedDevOrigins: ['192.168.43.26'],
-}
 
 };